requirements.txt
Flask==2.3.2
flask-cors==3.1.1
transformers==4.31.0
torch==2.1.0
vosk==0.3.30
ffmpeg-python==4.4.0
newspaper3k==0.2.8
nltk==3.8.1
textblob==0.15.3
werkzeug==2.3.7
